+++ projects/ausmittlung-lib/src/lib/components/proportional-election/proportional-election-ballot-content/proportional-election-ballot-content.component.ts
@@ -63,7 +63,7 @@ export class ProportionalElectionBallotContentComponent {
       return;
     }
 
-    this.ballotUiService.removeCandidateAtFirstFoundPosition(candidate, this.ballotUiData);
+    this.ballotUiService.removeCandidateAtLastFoundPosition(candidate, this.ballotUiData);
     this.contentChanged.emit();
   }