+++ projects/monitoring/src/app/app.component.ts
@@ -11,8 +11,9 @@ import { TranslateService } from '@ngx-translate/core';
 import { SnackbarService, ThemeService } from '@abraxas/voting-lib';
 import { LanguageService } from '../../../ausmittlung-lib/src/lib/services/language.service';
 import { LocationStrategy } from '@angular/common';
-import { Subscription } from 'rxjs';
+import { firstValueFrom, Subscription } from 'rxjs';
 import * as moment from 'moment';
+import { Title } from '@angular/platform-browser';
 
 @Component({
   selector: 'app-root',
@@ -23,6 +24,8 @@ export class AppComponent implements OnInit, OnDestroy {
   public hasTenant = false;
   public loading = false;
   public theme?: string;
+  public customLogo?: string;
+  public appTitle: string = '';
 
   @ViewChild('snackbar')
   public snackbarComponent?: SnackbarComponent;
@@ -39,6 +42,7 @@ export class AppComponent implements OnInit, OnDestroy {
     private readonly snackbarService: SnackbarService,
     private readonly languageService: LanguageService,
     private readonly locationStrategy: LocationStrategy,
+    private readonly title: Title,
   ) {
     // enable automatic silent refresh
     this.oauthService.setupAutomaticSilentRefresh();
@@ -54,9 +58,11 @@ export class AppComponent implements OnInit, OnDestroy {
     });
     this.subscriptions.push(snackbarSubscription);
 
-    // This prevents a short flickering of the default theme (if another theme has been set)
-    const themeSubscription = themeService.theme$.subscribe(theme => (this.theme = theme));
+    const themeSubscription = themeService.theme$.subscribe(theme => this.onThemeChange(theme));
     this.subscriptions.push(themeSubscription);
+
+    const logoSubscription = themeService.logo$.subscribe(logo => (this.customLogo = logo));
+    this.subscriptions.push(logoSubscription);
   }
 
   public async switchTenant(): Promise<void> {
@@ -99,5 +105,18 @@ export class AppComponent implements OnInit, OnDestroy {
     for (const subscription of this.subscriptions) {
       subscription.unsubscribe();
     }
+  }
+
+  private async onThemeChange(theme?: string): Promise<void> {
+    if (!theme) {
+      return;
+    }
+
+    // Cannot use translations.instant here, as the translations may not have been loaded yet
+    // It would then just display the non-translated string
+    this.appTitle = await firstValueFrom(this.translations.get('APP.TITLE.' + theme));
+    this.title.setTitle(this.appTitle);
+
+    this.theme = theme;
   }
 }