+++ projects/monitoring/src/app/components/end-result-page/end-result-page.component.ts
@@ -25,7 +25,4 @@ export class EndResultPageComponent {
 
   @Input()
   public domainOfInfluenceType?: DomainOfInfluenceType;
-
-  @Output()
-  public export: EventEmitter<void> = new EventEmitter<void>();
 }